pyproject.toml
[project]
name = "lattice"
version = "0.1.0"
description = "AI-powered political intelligence platform"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "psycopg2-binary>=2.9.9",
    "supabase>=2.0.0",
    "streamlit>=1.40.0",
    "pandas>=2.2.0",
    "scikit-learn>=1.5.0",
    "plotly>=5.24.0",
    "requests>=2.32.0",
    "python-dotenv>=1.0.0",
    "beautifulsoup4>=4.12.0",
    "numpy>=1.26.0",
    "pymupdf>=1.27.2.2",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
